Split Architecture
Much cleaner now, can focus on the maths better. Also added simple logic for is_first_layer, so now iterating over the list of layers and passing the LS to the upstream layer (layer with an index 1 lower), is completely fine, as the update_ls ignores None, and a layer returns None for the LS if it is the first layer. Didn't implement the derivative mathematics, just the architecture for now.