Clean and modular refactor
The split architecture idea was garbage. Now I tried to make this much closer to how pytorch functions, somewhat there. I want to make this into something clean and proper though, so still some way to go, but nonetheless it seems to work to learn. Not quite sure how to approach the softmax layer, or the sigmoid one that I add to plan (make a bernoulli distribution probability for each index in the output) and how to approach with the loss there, but still, much better now.