V 0.0.0
The most basic rudimentary version, it works. Given some input and a desired spike output, the model can learn, just check testing.py. Plotting spike trains is easy, getting the loss and inverse functions is as well, the idea of _leaky is to be used only with Sequential, other than that its largely pytorchesque